patch: fix error on startup
I noticed on the ticket 391 that the modified patch cannot be applied
properly on the last version:

  INFO: Patching Grocy to fix relative URL handling...
  patching file views/layout/default.blade.php
  Hunk #1 FAILED at 88.
  1 out of 1 hunk FAILED -- saving rejects to file views/layout/default.blade.php.rej

I simply tried to apply the patch on top of the latest v4.0.3 release
and fixed the conflicts. The context has changed in v4.0: 'ActiveNav'
variable has been replaced by 'View' in commit d16d976d0 ("Simplified
viewjs / active page handling") in Grocy repo.

I don't know if this will fix the recent issues that have been reported
(391, 392, 395) because I'm not using this addon yet: I was looking at
trying it and I wanted to fix this issue first.

Signed-off-by: Matthieu Baerts <matttbe@gmail.com>